refactor(web): decompose converter workspace
